Fix: various compile warnings
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@8872 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Strategic/LuaInitNPCs.cpp
@@ -2687,7 +2687,7 @@ void LuaHandleSectorLiberation( INT16 sSectorX, INT16 sSectorY, INT8 bSectorZ, B
 
 void LuaHandleInteractiveActionResult( INT16 sSectorX, INT16 sSectorY, INT8 bSectorZ, 
 									   INT32 sGridNo, UINT8 bLevel, UINT8 ubId,
-									   UINT8 usActionType, INT32 sLuaactionid, INT32 difficulty, UINT16 skill )
+									   UINT16 usActionType, INT32 sLuaactionid, INT32 difficulty, UINT16 skill )
 {
 	const char* filename = "scripts\\Overhead.lua";
