+++ Tactical/ArmsDealerInvInit.h
@@ -39,7 +39,7 @@ typedef struct
 
 INT8	GetDealersMaxItemAmount( UINT8 ubDealerID, UINT16 usItemIndex );
 
-DEALER_POSSIBLE_INV *GetPointerToDealersPossibleInventory( UINT8 ubArmsDealerID );
+DEALER_POSSIBLE_INV *GetPointerToDealersPossibleInventory( UINT16 ubArmsDealerID );
 
 UINT8 GetCurrentSuitabilityForItem( INT8 bArmsDealer, UINT16 usItemIndex );
 UINT8 ChanceOfItemTransaction( INT8 bArmsDealer, UINT16 usItemIndex, BOOLEAN fDealerSelling, BOOLEAN fUsed );
